requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
frozenlist==1.5.0
git-filter-repo==2.47.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.3.2
numpy==2.2.0
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
pycountry==24.6.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
serpapi==0.1.5
setuptools==78.1.0
six==1.17.0
svgwrite==1.4.3
Tree==0.2.4
twilio==9.5.1
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
